Register metrics globally
