+++ manager/pkg/cronjob/task/data_retention.go
@@ -9,7 +9,7 @@ import (
 	"github.com/go-co-op/gocron"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/metrics"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/monitoring"
 	"github.com/stolostron/multicluster-global-hub/pkg/database"
 	"github.com/stolostron/multicluster-global-hub/pkg/database/models"
 )
@@ -43,7 +43,7 @@ var (
 )
 
 func init() {
-	metrics.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(0)
+	monitoring.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(0)
 }
 
 func DataRetention(ctx context.Context, retentionMonth int, job gocron.Job) {
@@ -52,9 +52,9 @@ func DataRetention(ctx context.Context, retentionMonth int, job gocron.Job) {
 	var err error
 	defer func() {
 		if err != nil {
-			metrics.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(1)
+			monitoring.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(1)
 		} else {
-			metrics.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(0)
+			monitoring.GlobalHubCronJobGaugeVec.WithLabelValues(RetentionTaskName).Set(0)
 		}
 	}()
